Move defs to src
