+++ src/app/objects/factions/objects-by-faction/objects-by-faction.component.ts
@@ -1,9 +1,9 @@
-import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
-import { ActivatedRoute, ParamMap, Params } from '@angular/router';
+import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
 import { ReplaySubject } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { DB_ObjectRef_ByComponent } from '../../../cdclient';
-import { DESTRUCTIBLE_COMPONENT_ID } from '../../../defs/components';
+import { DESTRUCTIBLE_COMPONENT_ID } from '../../../../defs/components';
 import { LuJsonService } from '../../../services';
 
 @Component({
