+++ src/defs/api.ts
@@ -1,4 +1,4 @@
-import { DB_BehaviorTemplate } from "../cdclient";
+import { DB_BehaviorTemplate } from "../app/cdclient";
 
 export interface Behavior {
     template: DB_BehaviorTemplate,
